Conflicts:
	CPSC585/CPSC585.suo

+++ CPSC585/CPSC585/src/InputMapper.cpp
@@ -4,13 +4,13 @@
 #include "ForwardForceEvent.h"
 
 void InputMapper::updateRotation(double controllerInputX, double controllerInputY){
-	rotation = btQuaternion(controllerInputX, controllerInputY, 0, 0);
+	rotation = btQuaternion((float)controllerInputX, (float)controllerInputY, 0, 0);
 
 }
 
 void InputMapper::updateRotation(AnalogEvent *e){
 //	rotation = btQuaternion(e->getXVal(), e->getYVal(), 0, 0);	
-	rotation = btQuaternion(0, -(e->getXVal()), 0, 0);	
+	rotation = btQuaternion(0, -(float)(e->getXVal()), 0, 0);	
 	//rotation.normalize();
 	rotation /= 6000;
 	EventSystemHandler::getInstance()->emitEvent(new RotationEvent(rotation));
